requirements.txt
SQLAlchemy
psycopg2-binary
python-dotenv
FastAPI
Uvicorn
python-telegram-bot
pandas
aiofiles
openpyxl